Create TestingFileHiveMetastore
