+++ plugin/trino-geospatial/src/test/java/io/trino/plugin/geospatial/TestSpatialJoins.java
@@ -26,7 +26,7 @@ import java.io.File;
 import java.util.Optional;
 
 import static io.trino.SystemSessionProperties.SPATIAL_PARTITIONING_TABLE_NAME;
-import static io.trino.plugin.hive.metastore.file.FileHiveMetastore.createTestingFileHiveMetastore;
+import static io.trino.plugin.hive.metastore.file.TestingFileHiveMetastore.createTestingFileHiveMetastore;
 import static io.trino.testing.TestingSession.testSessionBuilder;
 import static java.lang.String.format;
